replaced the Headers -> Webhooks Cypher query by using the abstraction layer NodeManager
